+++ test/Cli/Func.E2ETests/func-start.Tests/TestsWithFixtures/NodeV4Tests.cs
@@ -131,118 +131,6 @@ namespace Func.E2ETests.func_start.Tests.TestsWithFixtures
         }
 
         [Fact]
-        public async Task Start_EmptyEnvVars_HandledAsExpected()
-        {
-            int port = ProcessHelper.GetAvailablePort();
-
-            // Create a temporary working directory
-            var tempDir = Path.Combine(_fixture.WorkingDirectory, "temp_empty_env");
-            Directory.CreateDirectory(tempDir);
-            CopyDirectoryHelpers.CopyDirectory(_fixture.WorkingDirectory, tempDir);
-
-            // Add empty setting
-            var funcSettingsResult = new FuncSettingsCommand(_fixture.FuncPath, _fixture.Log)
-                                    .WithWorkingDirectory(tempDir)
-                                    .Execute(new[] { "add", "emptySetting", "EMPTY_VALUE" });
-            funcSettingsResult.Should().ExitWith(0);
-
-            // Modify settings file to have empty value
-            string settingsPath = Path.Combine(tempDir, "local.settings.json");
-            string settingsContent = File.ReadAllText(settingsPath);
-            settingsContent = settingsContent.Replace("EMPTY_VALUE", "");
-            File.WriteAllText(settingsPath, settingsContent);
-
-            // Call func start
-            var funcStartCommand = new FuncStartCommand(_fixture.FuncPath, _fixture.Log, "Start_EmptyEnvVars_HandledAsExpected");
-
-            funcStartCommand.ProcessStartedHandler = async (process, fileWriter) =>
-            {
-                await ProcessHelper.ProcessStartedHandlerHelper(port, process, fileWriter);
-            };
-
-            var result = funcStartCommand
-                        .WithWorkingDirectory(tempDir)
-                        .Execute(new[] { "--port", port.ToString() });
-
-            // Validate function works and doesn't show skipping message
-            result.Should().NotHaveStdOutContaining("Skipping 'emptySetting' from local settings as it's already defined in current environment variables.");
-
-            // Clean up temporary directory
-            try { Directory.Delete(tempDir, true); } catch { }
-        }
-
-        [Fact]
-        public async Task Start_LanguageWorker_LogLevelOverridenViaSettings_LogLevelSetToExpectedValue()
-        {
-            int port = ProcessHelper.GetAvailablePort();
-
-            // Create a temporary working directory with debug log level
-            var tempDir = Path.Combine(_fixture.WorkingDirectory, "temp_log_level");
-            Directory.CreateDirectory(tempDir);
-            CopyDirectoryHelpers.CopyDirectory(_fixture.WorkingDirectory, tempDir);
-
-            // Add debug log level setting
-            var funcSettingsResult = new FuncSettingsCommand(_fixture.FuncPath, _fixture.Log)
-                                    .WithWorkingDirectory(tempDir)
-                                    .Execute(new[] { "add", "AzureFunctionsJobHost__logging__logLevel__Default", "Debug" });
-            funcSettingsResult.Should().ExitWith(0);
-
-            // Call func start
-            var funcStartCommand = new FuncStartCommand(_fixture.FuncPath, _fixture.Log, "Start_LanguageWorker_LogLevelOverridenViaSettings_LogLevelSetToExpectedValue");
-
-            funcStartCommand.ProcessStartedHandler = async (process, fileWriter) =>
-            {
-                await ProcessHelper.ProcessStartedHandlerHelper(port, process, fileWriter, "HttpTrigger?name=Test");
-            };
-
-            var result = funcStartCommand
-                        .WithWorkingDirectory(tempDir)
-                        .Execute(new[] { "--port", port.ToString(), "--verbose" });
-
-            // Validate we see detailed worker logs
-            result.Should().HaveStdOutContaining("Workers Directory set to");
-
-            // Clean up temporary directory
-            try { Directory.Delete(tempDir, true); } catch { }
-        }
-
-        [Fact]
-        public async Task Start_LanguageWorker_LogLevelOverridenViaHostJson_LogLevelSetToExpectedValue()
-        {
-            int port = ProcessHelper.GetAvailablePort();
-
-            // Create a temporary working directory with host.json log level
-            var tempDir = Path.Combine(_fixture.WorkingDirectory, "temp_host_log_level");
-            Directory.CreateDirectory(tempDir);
-            CopyDirectoryHelpers.CopyDirectory(_fixture.WorkingDirectory, tempDir);
-
-            // Modify host.json to set log level
-            string hostJsonPath = Path.Combine(tempDir, "host.json");
-            string hostJsonContent = "{\"version\": \"2.0\",\"logging\": {\"logLevel\": {\"Default\": \"None\"}}}";
-            File.WriteAllText(hostJsonPath, hostJsonContent);
-
-            // Call func start
-            var funcStartCommand = new FuncStartCommand(_fixture.FuncPath, _fixture.Log, "Start_LanguageWorker_LogLevelOverridenViaHostJson_LogLevelSetToExpectedValue");
-
-            funcStartCommand.ProcessStartedHandler = async (process, fileWriter) =>
-            {
-                await ProcessHelper.ProcessStartedHandlerHelper(port, process, fileWriter, "HttpTrigger?name=Test");
-            };
-
-            var result = funcStartCommand
-                        .WithWorkingDirectory(tempDir)
-                        .Execute(new[] { "--port", port.ToString() });
-
-            // Validate minimal worker logs due to "None" log level
-            result.Should().HaveStdOutContaining("Worker process started and initialized");
-            result.Should().NotHaveStdOutContaining("Initializing function HTTP routes");
-
-            // Clean up temporary directory
-            try { Directory.Delete(tempDir, true); } catch { }
-        }
-
-
-        [Fact]
         public async Task DontStart_InProc6_SpecifiedRuntime_ForNonDotnetApp()
         {
             int port = ProcessHelper.GetAvailablePort();
